Create a table for live content items
When we look at links that are missing from the publishing api but
are in rummager, we would like to know a bit more about what published
them.

This new table contains metadata about the content (not the links)
from the publishing api, so we can reference this in our queries.