Add scripts to show how links differ
bin/compare previously only showed link_types that we found in both
rummager and publishing api.

Now we also look at link types that are not present in publishing api at
all. In some cases (specialist publisher) this is because we haven't
migrated them yet. We may also have false positives because we are
not matching content properly.

./bin/compare now shows a breakdown by app of this missing data.

./bin/missing_app_links allows you to extract a list by app

./bin/mismatched_content shows the content we failed to match by base path