Match lead_organisations to organisations
At the moment some content has been tagged with a slightly different data model.

First way:
- Organisations is everyhing
- Lead organisations (if present) indicates which ones are lead vs supporting

Second way:
- No organisations
- Lead organisations is one link type
- Supporting organisations is another link type

:(

If we assume that either a lead organisations link or an organisation link means
the organisations are tagged, we can filter out some false positives and get at
what we really want to see, which is the content that has not been tagged in the
publishing api.