Start on rewriting for sqlite and publishing-api
We are moving to sqlite for our local db to avoid a dependency on postgres
on the jenkins node which will run the checker.

We are introducing a new endpoint for publishing-api to allow
us to query public apis rather than go to the publishing-api db.

This is the first pass of the necessary rewrite, establishing the skeleton for import
and checks, and getting the data import working with sqlite as it was before. We have changed
the imported table structure to better suit our needs and because sqlite3 does not support
the same data types as postgres.

We use gds-api-adapters where possible, and we will need to switch to using it for the
new endpoint when that endpoint is finished.