Add basic whitelisting functionality
We will need to suppress output from checks that we know about in order to keep
Jenkins green. For example, something we know is being fixed already, or something
for which we need to delay the fix.

If we decide to *never* fix something, we should filter it in the check rather than in
the whitelist. Therefore, every whitelist entry can and should have an expiry date and
a reason for existing. This allows us to report errors if any whitelist entry is expired.