Remove thread pools from importers
We've observed the checker starving its own import of publishing-api
content by sending many requests for the content_id lookup
during the rummager import.
This doesn't always happen but it is not worth the effort to investigate.
We've deployed to jenkins on integration, so we now care
more about consistency than speed.
Since the resources in integration are different from those on
our laptops, we shouldn't expect the execution model used
during dev to carry over particularly well.
We should also consider other users of the integration environment.