Ignore withdrawn items when checking what is missing from publishing-api
We cannot link rummager base paths to pubilshing-api content ids
when the item in question is withdrawn, because the lookup by
base path endpoint only returns published items.

We still need to include withdrawn items in the rummager import
since we need to prevent false positives in the opposite check.

So, this PR includes the is_withdrawn field from rummager and examines it
in the single check where it has to be used as a filter.

@davidslv and @benhyland