Sweep all the existing problems under the rug
There is a good chunk of content that has missing links in either rummager or publishing api.

The organisations do not have much of an impact, leave these till last.
The mainstream browse pages and topics can cause issues with browse, we should follow up on these separately.

This whitelists everything that is currently flagged. This should allow us to set up alerting.